# Conflicts:
#	fpa-config/src/main/resources/project.properties

+++ fpa-service/src/main/java/com/roof/fpa/cardslot/service/api/ICardSlotService.java
@@ -56,7 +56,7 @@ public interface ICardSlotService {
 	 */
 	public abstract List<CardSlotVo> selectForList(CardSlot cardSlot);
 
-	CardSlotVo selectBySceneIdAndNumb(Long sceneId, Integer numb);
+	CardSlotVo selectBySceneIdAndNumb(Long sceneId, Long numb);
 
 	/**
 	 * 按对象中的非空属性作为条件，进行分页查询
